refactor: copy resources and sidecars on the Cargo build script
