+++ tooling/bundler/Cargo.toml
@@ -22,8 +22,8 @@ exclude = [
 ]
 
 [dependencies]
+tauri-utils = { version = "1.0.0-beta.3", path = "../../core/tauri-utils", features = [ "resources" ] }
 ar = "0.9.0"
-glob = "0.3.0"
 icns = "0.3"
 image = "0.23.14"
 libflate = "1.1"
