+++ tooling/bundler/src/bundle/common.rs
@@ -7,7 +7,7 @@ use std::{
   ffi::OsStr,
   fs::{self, File},
   io::{self, BufWriter, Write},
-  path::{Component, Path, PathBuf},
+  path::Path,
   process::{Command, Stdio},
 };
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
@@ -131,23 +131,6 @@ pub fn copy_dir(from: &Path, to: &Path) -> crate::Result<()> {
     }
   }
   Ok(())
-}
-
-/// Given a path (absolute or relative) to a resource file, returns the
-/// relative path from the bundle resources directory where that resource
-/// should be stored.
-pub fn resource_relpath(path: &Path) -> PathBuf {
-  let mut dest = PathBuf::new();
-  for component in path.components() {
-    match component {
-      Component::Prefix(_) => {}
-      Component::RootDir => dest.push("_root_"),
-      Component::CurDir => {}
-      Component::ParentDir => dest.push("_up_"),
-      Component::Normal(string) => dest.push(string),
-    }
-  }
-  dest
 }
 
 /// Prints a message to stderr, in the same format that `cargo` uses,
