+++ tooling/bundler/src/bundle/macos/ios.rs
@@ -45,7 +45,7 @@ pub fn bundle_project(settings: &Settings) -> crate::Result<Vec<PathBuf>> {
 
   for src in settings.resource_files() {
     let src = src?;
-    let dest = bundle_dir.join(common::resource_relpath(&src));
+    let dest = bundle_dir.join(tauri_utils::resources::resource_relpath(&src));
     common::copy_file(&src, &dest)
       .with_context(|| format!("Failed to copy resource file {:?}", src))?;
   }
