+++ tooling/bundler/src/error.rs
@@ -9,15 +9,12 @@ use thiserror::Error as DeriveError;
 #[derive(Debug, DeriveError)]
 #[non_exhaustive]
 pub enum Error {
+  /// Error running tauri_utils API.
+  #[error("{0}")]
+  Resource(#[from] tauri_utils::Error),
   /// Bundler error.
   #[error("{0}")]
   BundlerError(#[from] anyhow::Error),
-  /// Failed to use glob pattern.
-  #[error("`{0}`")]
-  GlobError(#[from] glob::GlobError),
-  /// Invalid glob pattern.
-  #[error("`{0}`")]
-  GlobPatternError(#[from] glob::PatternError),
   /// I/O error.
   #[error("`{0}`")]
   IoError(#[from] io::Error),
