+++ tooling/cli.rs/src/dev.rs
@@ -74,28 +74,6 @@ pub fn command(options: Options) -> Result<()> {
   };
   let config = get_config(merge_config.as_deref())?;
 
-  let (settings, out_dir) = {
-    let config_guard = config.lock().unwrap();
-    let config_ = config_guard.as_ref().unwrap();
-    let app_settings = crate::interface::rust::AppSettings::new(config_)?;
-    let out_dir = app_settings
-      .get_out_dir(options.target.clone(), true)
-      .with_context(|| "failed to get project out directory")?;
-    let settings = crate::interface::get_bundler_settings(
-      app_settings,
-      options.target.clone(),
-      &Default::default(),
-      config_,
-      &out_dir,
-      false,
-      None,
-    )
-    .with_context(|| "failed to build bundler settings")?;
-    (settings, out_dir)
-  };
-  settings.copy_resources(&out_dir)?;
-  settings.copy_binaries(&out_dir)?;
-
   if let Some(before_dev) = &config
     .lock()
     .unwrap()
